+++ src/components/Greetings/Welcome.stories.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { themeDecorator } from "../../storybookUtils";
+import { Welcome } from "./Welcome";
+
+export const Standard: React.FC = () => <Welcome />;
+
+export default {
+  title: "Components/Welcome",
+  decorators: [themeDecorator()],
+  component: Welcome,
+};
